fix(activerecord/encryption): extract encoding helpers, wire onConfigure cache invalidation, rename compress param
- Extract normalizeEncoding + replaceUnencodable to encoding-helpers.ts;
  both Encryptor and EncryptedAttributeType now share a single impl
- Encryptor registers Configurable.onConfigure to clear _defaultKeyProviderCache
  on key rotation / config changes
- Rename compressIfWorthIt param 'string' → 'clearText'
